requirements.txt
Django==1.6.1
MySQL-python==1.2.5
South==0.8.4
beautifulsoup4==4.3.2
django-tastypie==0.11.0
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.2.1
six==1.5.2
wsgiref==0.1.2
